Nicolas Regnault: - fix a bug in TightBindingModelTwoOrbitalSquareLattice - add a few options to TIEntanglementEntropyRealSpacePartition
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3849 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FTI/src/Tools/FTITightBinding/TightBindingModelTwoOrbitalSquareLattice.cc
@@ -53,7 +53,7 @@ using std::endl;
 // gammaX = boundary condition twisting angle along x
 // gammaY = boundary condition twisting angle along y
 // storeOneBodyMatrices = flag to indicate if the one body transformation matrices have to be computed and stored
-TightBindingModelTwoOrbitalSquareLattice::TightBindingModelTwoOrbitalSquareLattice(int nbrSiteX, int nbrSiteY, int t1, int t2, int t3, int foldingFactor, double mus, double gammaX, double gammaY,
+TightBindingModelTwoOrbitalSquareLattice::TightBindingModelTwoOrbitalSquareLattice(int nbrSiteX, int nbrSiteY, double t1, double t2, double t3, int foldingFactor, double mus, double gammaX, double gammaY,
         AbstractArchitecture* architecture, bool storeOneBodyMatrices)
 {
     this->NbrSiteX = nbrSiteX;
